==> step_1_data.ipynb <==
# !pip install -r requirement.txt

from google.cloud import bigquery
import pandas as pd
import os 

# !pip install --upgrade google-cloud-storage
# !export GOOGLE_APPLICATION_CREDENTIALS="/Users/neil/Desktop/Hui_Yuan/Projects/Global AI/GA_hui/apikey.json"
# test environment and configuration done

# from google.cloud import storage
# def implicit(): 
#     # If you don't specify credentials when constructing the client, the
#     # client library will look for credentials in the environment.
#     storage_client = storage.Client()

#     # Make an authenticated API request
#     buckets = list(storage_client.list_buckets())
#     print(buckets)
# implicit()


rawQuery = client.query("""
SELECT DATE,THEMES,DocumentIdentifier \
FROM `gdelt-bq.gdeltv2.gkg` \
WHERE DATE>=20181231000000 AND DATE<20190107000000 \
AND Locations like "%US%" \
AND THEMES like "%ENV_GREEN%" """)

raw = rawQuery.to_dataframe()
raw.head()

if not os.path.isdir('./data/'):
    os.makedirs('/Users/neil/Desktop/Hui_Yuan/Projects/Global AI/GA_hui/data')
raw.to_csv('./data/bigquery_raw.csv')

## download gdelt done
