==> step_3_topic.ipynb <==
import pandas as pd
import numpy as np
import sys
sys.path.append('./lib/')
from utility import get_text,ProcessPipeline
### Part 2 Package
from gensim import corpora, models
from IPython.display import Image
import pickle

import matplotlib.pyplot as plt
%matplotlib inline

### read pickle file
with open('data/news.pickle', 'rb') as handle:
    texts = pickle.load(handle)

pipeline = ProcessPipeline(texts)

textsProcessed = pipeline.run()

# double check
textsProcessed[0][:10]

### model topic (LDA)

from gensim import corpora, models
import gensim
from gensim.models import CoherenceModel

"""
Create a dictionary from ‘textsProcessed’ containing the number of times a word appears in the training set.
The Dictionary() function traverses texts, assigning a unique integer id to each unique token while also collecting word counts and relevant statistics. 
    To see each token’s unique integer id, try print(dictionary.token2id).
""" 
dictionary = corpora.Dictionary(textsProcessed)
print(dictionary.token2id)

# transform to bag-of-words ( first element is word ind, second one is number)
corpus = [dictionary.doc2bow(text) for text in textsProcessed]
corpus

ldamodel = gensim.models.ldamodel.LdaModel(corpus, num_topics=3, id2word = dictionary, passes=20)

ldamodel.print_topics(num_topics=3, num_words=10)

# Evaluate LDA model based on Perplexity & Coherece Score
from gensim.models import CoherenceModel
# Compute Perplexity
print('\nPerplexity: ', ldamodel.log_perplexity(corpus))  # a measure of how good the model is. lower the better.

# Compute Coherence Score
coherence_model_lda = CoherenceModel(model=ldamodel, texts=textsProcessed, dictionary=dictionary, coherence='c_v')
coherence_lda = coherence_model_lda.get_coherence()
print('\nCoherence Score: ', coherence_lda)

# Check each document and its topics
for i in range(len(corpus)):
    tmpTopic = ldamodel[corpus[i]]
    print(f"For doc {i}, corresponding topics & probability are {tmpTopic}")

# Visualize the topics
from pyLDAvis import gensim as pyLDAvis_gensim
import pyLDAvis
pyLDAvis.enable_notebook()
vis = pyLDAvis_gensim.prepare(ldamodel, corpus, dictionary)
vis

# find the optimal topic
def compute_coherence_values(dictionary, corpus, texts, limit, start=2, step=3):
    """
    Func:
        Compute c_v coherence for various number of topics
    Input:
        dictionary : Gensim dictionary
        corpus : Gensim corpus
        texts : List of input texts
        limit : Max num of topics
    Returns:
        model_list : List of LDA topic models
        coherence_values : Coherence values corresponding to the LDA model with respective number of topics
    """
    coherence_values = []
    model_list = []
    for num_topics in range(start, limit, step):
        model = gensim.models.ldamodel.LdaModel(corpus, num_topics=num_topics, id2word = dictionary)
#         model = gensim.models.wrappers.LdaMallet(mallet_path, corpus=corpus, num_topics=num_topics, id2word=id2word)
        model_list.append(model)
        coherencemodel = CoherenceModel(model=model, texts=texts, dictionary=dictionary, coherence='c_v')
        coherence_values.append(coherencemodel.get_coherence())

    return model_list, coherence_values

model_list, coherence_values = compute_coherence_values(dictionary=dictionary, corpus=corpus, texts=textsProcessed, start=2, limit=40, step=6)

coherence_values

from gensim import corpora, models
import gensim
from gensim.models import CoherenceModel
from pyLDAvis import gensim as pyLDAvis_gensim
import pyLDAvis

class LDA:
    def __init__(self,texts,num_topics=3):
        ### input texts: list of texts eg: [["jason","while"],["bay","wheel"]]
        self.texts = texts
        self.dictionary = corpora.Dictionary(texts)
        self.corpus = [self.dictionary.doc2bow(text) for text in texts]
        
        self.ldamodel = gensim.models.ldamodel.LdaModel(self.corpus, num_topics=num_topics, id2word = self.dictionary, passes=20)
        
    def get_topics(self,num_topics=3,num_words=5):
        topics = self.ldamodel.print_topics(num_topics=num_topics, num_words=num_words)
        print(f"topics: {topics}")
    def get_doc_topics(self,docId=0,print_topics=False):
        tmpTopic = self.ldamodel[self.corpus[docId]]
        if print_topics: print(f"For doc {docId}, corresponding topics & probability are {tmpTopic}")
        return tmpTopic
    
    def get_perplexity(self):
        return self.ldamodel.log_perplexity(self.corpus)  # a measure of how good the model is. lower the better.

    def get_coherence(self):
        # Compute Coherence Score
        coherence_model_lda = CoherenceModel(model=self.ldamodel, texts=self.texts, dictionary=self.dictionary, coherence='c_v')
        coherence_lda = coherence_model_lda.get_coherence()
        print('\nCoherence Score: ', coherence_lda)
        return self.coherence_lda

    def plot(self):
        # Visualize the topics
        pyLDAvis.enable_notebook()
        vis = pyLDAvis_gensim.prepare(self.ldamodel,self.corpus, self.dictionary)
        return vis

### done
