==> step_4_sentiment.ipynb <==
import pandas as pd
import numpy as np
from lib.utility import get_text, ProcessPipeline
import pickle
from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
from textblob import TextBlob

### read pickle file for get strings rather than words
with open('./data/news.pickle', 'rb') as handle:
    texts = pickle.load(handle)

pipeline = ProcessPipeline(texts,steps=['langdetection','summarization'])
textsSummarized = pipeline.run(return_str=True)

### save as pickle
with open('./data/text_for_sentiment.pickle', 'wb') as handle:
    pickle.dump(textsSummarized, handle, protocol=pickle.HIGHEST_PROTOCOL)

### vaderSentiment

# use pretrained vader sentiment model
analyzer = SentimentIntensityAnalyzer()
# eg:
eg = textsSummarized[0]
senti_eg = analyzer.polarity_scores(eg)
print(senti_eg)

# for the whole texts
senti_list = list(map(lambda x:analyzer.polarity_scores(x),textsSummarized))
print(senti_list[:5])


# In general, we use compound score as sentiment scores
senti_list_compound = list(map(lambda x:x['compound'],senti_list))
print(senti_list_compound[:5])

# do classification
senti_class = list(map(lambda x:'Positive' if x>0 else ('Negative' if x<0  else 'Neutral'),senti_list_compound))
senti_class[:5]

# save to csv
raw_0 = pd.read_csv('./data/bigquery_raw.csv',index_col=0)
raw_0['Sentiment_score'] = senti_list_compound
raw_0['Sentiment'] = senti_class
print(raw_0.head(5))
raw_0.to_csv('./data/news_with_senti.csv')

# To figure out most negative or positive news
def most_senti_news(df,Pos = True,top = 10):
    """
    default to find most positive news by using compound score
    return two seperate lists of THEMES and DocumentIdentifier
    """
    if Pos:
        new = df.sort_values(by=['Sentiment_score'], ascending=False).reset_index(drop=True)
    else:
        new = df.sort_values(by=['Sentiment_score'], ascending=True).reset_index(drop=True)
    return new['THEMES'][:top], new['DocumentIdentifier'][:top]


### TextBlob

blob_eg = TextBlob(eg).sentiment
print("blob_eg:",blob_eg)

# we could train a supervised classifier for this projects 
# set an example from the document of textblob
# use textblob classifier to easily understand for finance team
from textblob.classifiers import NaiveBayesClassifier
train = [
     ('I love this sandwich.', 'pos'),
     ('this is an amazing place!', 'pos'),
     ('I feel very good about these beers.', 'pos'),
     ('this is my best work.', 'pos'),
     ("what an awesome view", 'pos'),
     ('I do not like this restaurant', 'neg'),
     ('I am tired of this stuff.', 'neg'),
     ("I can't deal with this", 'neg'),
     ('he is my sworn enemy!', 'neg'),
     ('my boss is horrible.', 'neg')
 ]
test = [
     ('the beer was good.', 'pos'),
     ('I do not enjoy my job', 'neg'),
     ("I ain't feeling dandy today.", 'neg'),
     ("I feel amazing!", 'pos'),
     ('Gary is a friend of mine.', 'pos'),
    ("I can't believe I'm doing this.", 'neg')
 ]
cl = NaiveBayesClassifier(train)
blob = list(map(lambda x: cl.classify(x[0]),test))
blob

### Word embedding


# train a model by our docs
# but we need the finance team's support to get labels
# set an example to do, incited from document of tf-idf

# TF-IDF
from sklearn.feature_extraction.text import TfidfVectorizer
texts = [
    "The cat sat on the hat",
    "The dog ate the cat and the hat",
    "Dog and cat are good friends"
]
tfidf = TfidfVectorizer()
features = tfidf.fit_transform(texts)
print("tf_idf:\n\n",pd.DataFrame(features.todense(),columns=tfidf.get_feature_names()))

# svd
uniqueTf = tfidf.get_feature_names()
coMatrix = np.random.randint(3, size=(5, len(uniqueTf)))
print('svd\n\n',pd.DataFrame(columns=uniqueTf,data=coMatrix))
u, s, vt  = np.linalg.svd(coMatrix,full_matrices=False)
print(f"shape of SVD component:\n\n {u.shape},{s.shape},{vt.shape}")
print('score:\n\n',pd.DataFrame(u@np.diag(s)@vt).head(2))

### Modulize

class sentiment:
    def __init__(self, texts, df = None,Pos = True,top = 10, path):
        self.texts = texts
        self.df = df
        self.Pos = Pos
        self.top = top
        self.path = path
    
    def get_data(self,texts):
        pipeline = ProcessPipeline(texts,steps=['langdetection','summarization'])
        textsSummarized = pipeline.run(return_str=True)
        return textsSummarized
    
    def most_senti_news(self,df,Pos,top):
        """
        default to find most positive news by using compound score
        return two seperate lists of THEMES and DocumentIdentifier
        """
        if Pos:
            new = df.sort_values(by=['Sentiment_score'], ascending=False).reset_index(drop=True)
        else:
            new = df.sort_values(by=['Sentiment_score'], ascending=True).reset_index(drop=True)
        return new['THEMES'][:top], new['DocumentIdentifier'][:top]

    def senti_vendar(self,texts,df):
        textsSummarized = self.get_data(texts)
        analyzer = SentimentIntensityAnalyzer()
        senti_list = list(map(lambda x:analyzer.polarity_scores(x),textsSummarized))
        senti_list_compound = list(map(lambda x:x['compound'],senti_list))
        senti_class = list(map(lambda x:'Positive' if x>0 else ('Negative' if x<0  else 'Neutral'),senti_list_compound))
        df['Sentiment_score'] = senti_list_compound
        df['Sentiment'] = senti_class
        df.to_csv(path,index_col = 0)
        return self.df
    
    def run(self,workers=6):
        with ProcessPoolExecutor(max_workers=workers) as executor:
            res = executor.map(self.senti_vendar(), self.texts, self.df)         
        return list(res)
