==> step_5_tagging.ipynb <==
import pandas as pd
import numpy as np
from lib.utility import get_text, ProcessPipeline
import nltk
from nltk.tokenize import word_tokenize
from nltk.tag import pos_tag
import pickle
from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor

### Part 1: get raw text data 

### read pickle file
with open('./data/news.pickle', 'rb') as handle:
    texts = pickle.load(handle) 

pipeline = ProcessPipeline(texts)

textsProcessed = pipeline.run(return_str=False)

textsProcessed[0][:10]

len(textsProcessed)

### Part 2: Sequence Tagging

### 2.1 NLTK sequence tagging

### Pos: part of speech

print(pos_tag(textsProcessed[0]))

ex = 'European authorities fined Google a record $5.1 billion on Wednesday for abusing its power in the mobile phone market and ordered the company to alter its practices'

# Token: list of word: sequence tagging

nltk.word_tokenize(ex)

def tagSec(sent):
    sent = nltk.word_tokenize(sent)
    sent = nltk.pos_tag(sent)
    return sent

print(tagSec(ex))

### Explain of tags
nltk.help.upenn_tagset()

## 2.2 Spacy nlp: Tagging, entity recognition

!python -m spacy download en <br>
Tutorial: https://spacy.io/models/

from pprint import pprint

import spacy
from spacy import displacy
from collections import Counter
import en_core_web_sm

nlp = spacy.load('en')

pipeline = ProcessPipeline()
textsSummarized = list(map(pipeline.summarize,texts))

eg = textsSummarized[0]
eg

#### Token api

doc = nlp(eg)

doc

for token in doc:
    print((token.text,token.lemma_,token.is_stop,token.pos_,token.tag_))

#### Entity api

for X in doc.ents:
    print((X.text, X.label_))

#### Sentence api

### Show doc sentences
sentences = [sen for sen in  doc.sents]
sentences

#### Visualization

### Display sentence relationship with entity style
displacy.render(nlp(str(sentences[0])), jupyter=True, style='ent')

### Display sentence relationship
displacy.render(nlp(str(sentences[0])), style='dep', jupyter = True, options = {'distance': 100})

### Display whole doc 
displacy.render(doc,jupyter=True,style='ent')

###  Modulize 

import spacy
from spacy import displacy
from collections import Counter
import en_core_web_sm
nlp = spacy.load('en')

class entityPipeline:
    def __init__(self,texts,targetPOSs=['NOUN']):
        ### texts is list of strings
        self.texts = texts
        self.targetPOSs = targetPOSs

    def run(self,n_threads=16):
        nlp = spacy.load('en')

        res = []
        for doc in nlp.pipe(self.texts, n_threads=n_threads, batch_size=10000):
            featureDic = {}
            entDic = self.extract_entity(doc)
            POSDic = self.extract_POS(doc,self.targetPOSs)
            featureDic.update(entDic)
            featureDic.update(POSDic)
            res.append(featureDic)
        return res
        
    def process(self,text):
        ### text is string type
        nlp = spacy.load('en')
        doc = nlp(text)
        featureDic = {}
        entDic = self.extract_entity(doc)
        POSDic = self.extract_POS(doc,self.targetPOSs)
        featureDic.update(entDic)
        featureDic.update(POSDic)
        return featureDic
        
    def extract_entity(self,doc):
        entDic = {}
        for ent in doc.ents:
            if ent.text != "\n" and ent.label_ not in  ["CARDINAL",'DATE',"LAW"]:
                if ent.text in entDic:
                    entDic[ent.text]['count'] = entDic[ent.text]['count'] + 1  
                else:
                    tmpDic = {ent.text:{"count":1,"label":ent.label_}}
                    entDic.update(tmpDic)
        return entDic
    
    def extract_POS(self,doc,targetPOSs=["VERB"]):
        POSDic = {}
        for token in doc:
            if token.is_stop == False and token.pos_ in targetPOSs :
                if token.lemma_ in POSDic:
                    POSDic[token.lemma_]['count'] = POSDic[token.lemma_]['count'] + 1  
                else:
                    tmpDic = {token.lemma_:{"count":1,"POS":token.pos_}}
                    POSDic.update(tmpDic)
        return POSDic
